Rename storybook components folder
# Conflicts:
#	packages/tracking/package.json
